별점 추가 및 협업필터링 추가
협업 필터링 사용으로 별점 시스템 추가
주어진 게시판(boardId)과 유사한 다른 게시판들을 추천 목록으로 가져오는 기능

+++ src/main/java/com/today/demo/repository/ImagesRepository.java
@@ -12,5 +12,5 @@ import java.util.List;
 public interface ImagesRepository extends JpaRepository<Images, Long> {
 
 
-    List<Images> findByBoardId(int boardId);
+    List<Images> findByBoardId(long boardId);
 }
